requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
cloudinary==1.29.0
dj-database-url==1.0.0
Django==4.1.1
django-cloudinary-storage==0.3.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
